Maximum Likelihood estimator for a Cauchy model

Find the maximum likelihood, using numerical optimization with optim.

cauphylm(
  formula,
  data = list(),
  phy,
  model = c("cauchy", "lambda"),
  lower.bound = list(disp = 0, lambda = 0),
  upper.bound = list(disp = Inf, lambda = NULL),
  starting.value = list(disp = NULL, lambda = NULL),
  hessian = FALSE
)

Arguments

formula

a model formula.

data

a data frame containing variables in the model. If not found in data, the variables are taken from current environment.

phy

a phylogenetic tree of class phylo.

model

a model for the trait evolution. Only "cauchy" is allowed.

lower.bound

named list with lower bound values for the parameters.

upper.bound

named list with upper bound values for the parameters. See Details.

starting.value

starting value for the parameters. This should be a named list, with disp the dispersion parameter. The default initial values are computed from standard statistics used on (independent) Cauchy variables, see details.

hessian

if TRUE, then the numerical hessian is computed, for confidence interval computations. See compute_vcov.

Value

coefficients

the named vector of coefficients.

disp

the maximum likelihood estimate of the dispersion parameter.

logLik

the maximum of the log likelihood.

p

the number of all parameters of the model.

aic

AIC value of the model.

fitted.values

fitted values

residuals

raw residuals

y

response

X

design matrix

n

number of observations (tips in the tree)

d

number of dependent variables

formula

the model formula

call

the original call to the function

model

the phylogenetic model for the covariance

Details

Unless specified by the user, the initial values for the parameters are taken according to the following heuristics:

mu is the trimmed mean of the trait, keeping only 24% of the observations, as advocated in Rothenberg et al. 1964 (for models with the intercept only)

coef are obtained from a robust regression using lmrob.S (for linear models with several predictors)

disp is initialized from the trait centered and normalized by tip heights, with one of the following statistics, taken from Rousseeuw & Croux 1993:

IQR half of the inter-quartile range (see IQR);

MAD median absolute deviation with constant equal to 1 (see mad);

Sn Sn statistics with constant 0.7071 (see Sn);

Qn Qn statistics with constant 1.2071 (see Qn);

For model=lambda, the maximum default value is computed using phytools:::maxLambda, and is the ratio between the maximum height of a tip node over the maximum height of an internal node. This can be larger than 1.

References

Rothenberg T. J., Fisher F. M., Tilanus C. B. 1964. A Note on Estimation from a Cauchy Sample. Journal of the American Statistical Association. 59:460–463.

See also