Maximum Likelihood estimator for a Cauchy model

Find the maximum likelihood, using numerical optimization with optim.

fitCauchy(
  phy,
  trait,
  model = c("cauchy", "lambda"),
  method = c("reml", "random.root", "fixed.root"),
  starting.value = list(mu = NULL, disp = NULL, lambda = NULL),
  lower.bound = list(disp = 0, lambda = 0),
  upper.bound = list(disp = Inf, lambda = NULL),
  root.edge = 100,
  hessian = FALSE,
  optim = c("local", "global"),
  method.init.disp = c("Qn", "Sn", "MAD", "IQR")
)

Arguments

phy

a phylogenetic tree of class phylo.

trait

named vector of traits at the tips.

model

a model for the trait evolution. Only "cauchy" is allowed.

method

the method used to fit the process. One of reml (the default), random.root or fixed.root. See details.

starting.value

starting value for the parameters of the Cauchy. This should be a named list, with mu and disp the ancestral central and dispersion parameters. The default initial values are computed from standard statistics used on (independent) Cauchy variables, see details.

lower.bound

named list with lower bound values for the parameters.

upper.bound

named list with upper bound values for the parameters. See Details.

root.edge

multiplicative factor for the root dispersion, equal to the length of the root edge. Ignored if method!=random.root.

hessian

if TRUE, then the numerical hessian is computed, for confidence interval computations. See compute_vcov.

optim

if "local", only a local optimization around the initial parameter values is performed (the default). If "global", a global maximization is attempted using the "MLSL" approach (see nloptr).

method.init.disp

the initialization method for the dispersion. One of "Qn", "Sn", "MAD", "IQR". Default to the "Qn" statistics. See Details.

Value

A list, with the maximum likelihood rate parameter, and the likelihood value.

Details

Unless specified by the user, the initial values for the parameters are taken according to the following heuristics:

mu is the trimmed mean of the trait, keeping only 24% of the observations, as advocated in Rothenberg et al. 1964 (for method="fixed.root")

disp is initialized from the trait centered and normalized by tip heights, with one of the following statistics, taken from Rousseeuw & Croux 1993:

IQR half of the inter-quartile range (see IQR);

MAD median absolute deviation with constant equal to 1 (see mad);

Sn Sn statistics with constant 0.7071 (see Sn);

Qn Qn statistics with constant 1.2071 (see Qn);

For model=lambda, the maximum default value is computed using phytools:::maxLambda, and is the ratio between the maximum height of a tip node over the maximum height of an internal node. This can be larger than 1.

The method argument specifies the method used for the fit:

method=reml the dispersion parameter is fitted using the REML criterion, obtained by re-rooting the tree to one of the tips.

method=random.root the root value is assumed to be a random Cauchy variable, centered at mu=0, and with a dispersion disp_root = disp * root.edge.

method=fixed.root the model is fitted conditionally on the root value mu, i.e. with a model where the root value is fixed and inferred from the data.

In the first two cases, the optimization is done on the dispersion only, while in the last case the optimization is on the root value and the dispersion.

References

Rothenberg T. J., Fisher F. M., Tilanus C. B. 1964. A Note on Estimation from a Cauchy Sample. Journal of the American Statistical Association. 59:460–463. Rousseeuw P.J., Croux C. 1993. Alternatives to the Median Absolute Deviation. Journal of the American Statistical Association. 88:1273–1283.

See also

Examples

# Simulate tree and data
phy <- ape::rphylo(5, 0.1, 0)
dat <- rTraitCauchy(n = 1, phy = phy, model = "cauchy", parameters = list(root.value = 0, disp = 1))
# Fit the data
fit <- fitCauchy(phy, dat, model = "cauchy", method = "reml")
fit
#> Call:
#> fitCauchy(phy = phy, trait = dat, model = "cauchy", method = "reml")
#> 
#>                 AIC logLik (restricted) 
#>               31.49              -14.74 
#> 
#> Parameter estimate(s) using REML:
#> dispersion: 0.989253