Maximum Likelihood estimator for a Cauchy model

Find the maximum likelihood, using numerical optimization with optim.

fitCauchy.internal(
  phy,
  X,
  y,
  model = c("cauchy", "lambda"),
  method = c("reml", "random.root", "fixed.root"),
  starting.value = list(x0 = NULL, disp = NULL, lambda = NULL),
  lower.bound = list(disp = 0, lambda = 0),
  upper.bound = list(disp = Inf, lambda = 1),
  root.edge = 100,
  optim = c("local", "global"),
  method.init.disp = "Qn",
  ...
)

Arguments

phy

a phylogenetic tree of class phylo.

model

a model for the trait evolution. One of "cauchy" or "lambda" (see Details).

method

the method used to fit the process. One of reml (the default), fixed.root or random.root. See Details.

starting.value

starting value for the parameters of the Cauchy. This should be a named list, with x0 and disp the root starting value and the dispersion parameter. The default initial values are computed from standard statistics used on (independent) Cauchy variables, see Details.

lower.bound

named list with lower bound values for the parameters. See Details for the default values.

upper.bound

named list with upper bound values for the parameters. See Details for the default values.

root.edge

multiplicative factor for the root dispersion, equal to the length of the root edge. Ignored if method!=random.root.

optim

if "local", only a local optimization around the initial parameter values is performed (the default). If "global", a global maximization is attempted using the "MLSL" approach (see nloptr).

method.init.disp

the initialization method for the dispersion. One of "Qn", "Sn", "MAD", "IQR". Default to the "Qn" statistics. See Details.

Value

A list, with the maximum likelihood rate parameter, and the likelihood value.

References

Rothenberg T. J., Fisher F. M., Tilanus C. B. 1964. A Note on Estimation from a Cauchy Sample. Journal of the American Statistical Association. 59:460–463.

See also